+++ __init__.py
@@ -38,7 +38,7 @@ bl_info = {
 # ==================================================================================================================
 
 
-from .BioBlender2 import *
+from .BioBlender import *
 
 
 # ==================================================================================================================
